Refactoring to support easier syslog format implementations
