+++ src/append/syslog/plain/mod.rs
@@ -1,23 +1,21 @@
-//! Implementation of syslog messages format with only a PRI part.
+//! Implementation of syslog messages format with only the PRI part of the header.
 
 use log::LogRecord;
 
-use append::syslog::severity;
-
-const FACILITY: u8 = 1 << 3; // USER facility
+use append::syslog::consts::{Facility, level_to_severity};
 
 /// Syslog message format that only adds the priority header to the message.
 #[derive(Debug)]
 pub struct Format;
 
 impl Format {
-    /// Creates new `Format` object. 
+    /// Creates new `Format` object.
     pub fn new() -> Format {
         Format{}
     }
-    /// Creates RFC 5424 message for the given log record
+    /// Creates a plain syslog message for the given log record.
     pub fn apply(&self, rec: &LogRecord) -> String {
-    	let priority = FACILITY | severity::level_to_severity(rec.level());
+    	let priority = Facility::USER as u8 | level_to_severity(rec.level());
     	format!("<{}> {}\n",
     	    priority,
     	    rec.args()
